+++ ecs.php
@@ -3,10 +3,10 @@
 declare(strict_types=1);
 
 use craft\ecs\SetList;
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Symplify\EasyCodingStandard\Config\ECSConfig;
 use Symplify\EasyCodingStandard\ValueObject\Option;
 
-return static function(ContainerConfigurator $containerConfigurator): void {
+return static function(ECSConfig $containerConfigurator): void {
     $parameters = $containerConfigurator->parameters();
     $parameters->set(Option::PARALLEL, true);
     $parameters->set(Option::PATHS, [