+++ src/adapters/BaseNewsletterAdapter.php
@@ -1,17 +1,13 @@
 <?php
-/**
- * @link https://www.simplonprod.co
- * @copyright Copyright (c) 2021 Simplon.Prod
- */
 
-namespace simplonprod\newsletter\adapters;
+namespace juban\newsletter\adapters;
 
 use craft\base\ConfigurableComponent;
 
 /**
  * BaseNewsletterAdapter class
  *
- * @author albanjubert
+ * @author juban
  **/
 abstract class BaseNewsletterAdapter extends ConfigurableComponent implements NewsletterAdapterInterface
 {