+++ src/config.php
@@ -3,9 +3,6 @@
  * Newsletter plugin for Craft CMS 3.x
  *
  * Craft CMS Newsletter plugin
- *
- * @link      https://www.simplonprod.co
- * @copyright Copyright (c) 2021 Simplon.Prod
  */
 
 /**
@@ -23,7 +20,7 @@
  */
 
 return [
-    "adapterType" => \simplonprod\newsletter\adapters\Mailjet::class,
+    "adapterType" => \juban\newsletter\adapters\Mailjet::class,
     "adapterTypeSettings" => [
         'apiKey' => '',
         'apiSecret' => '',