+++ src/models/NewsletterForm.php
@@ -1,19 +1,15 @@
 <?php
-/**
- * @link https://www.simplonprod.co
- * @copyright Copyright (c) 2021 Simplon.Prod
- */
 
-namespace simplonprod\newsletter\models;
+namespace juban\newsletter\models;
 
 use Craft;
 use craft\base\Model;
-use simplonprod\newsletter\Newsletter;
+use juban\newsletter\Newsletter;
 
 /**
  * NewsletterForm class
  *
- * @author albanjubert
+ * @author juban
  **/
 class NewsletterForm extends Model
 {